Improved motion model, always keep horizon horizontal. Azimuth is interpolated in world coordinates, pitch in local coordinates.
ToDo: There is still some unexplained translation when the view is rotated.
This is probably to do with how the SceneView is constructed from a camera, a pivot and a rotation.
We can't move the camera as this is overwritten by unity.
We can only set the pivot position and the rotation.